requirements.txt
asgi-redis==1.4.2
asgiref==1.1.2
attrs==17.2.0
autobahn==17.6.1
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2017.4.17
channels==1.1.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cycler==0.10.0
daphne==1.3.0
defusedxml==0.5.0
dj-static==0.0.6
Django==1.11.2
django-allauth==0.34.0
django-ckeditor==5.3.0
django-colorfield==0.1.14
django-cors-headers==2.1.0
django-debug-toolbar==1.8
django-easy-select2==1.3.4
django-filter==1.1.0
django-geoposition==0.3.0
django-js-asset==0.1.1
django-map-widgets==0.1.9
django-material==1.0.0
django-modelcluster==3.1
django-rest-auth==0.9.1
django-smart-selects==1.5.2
django-taggit==0.22.1
django-treasuremap==0.2.5
django-treebeard==4.1.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
dlib==19.15
drf-extensions==0.3.1
face-recognition==1.2.3
face-recognition-models==0.3.0
gunicorn==19.7.1
html5lib==0.999999999
humanize==0.5.1
hyperlink==17.2.1
idna==2.5
imageio==2.3.0
incremental==17.5.0
kiwisolver==1.0.1
matplotlib==2.2.3
msgpack-python==0.4.8
numpy==1.15.1
oauthlib==2.0.2
olefile==0.44
opencv-python==3.4.2.17
Pillow==5.2.0
pyparsing==2.2.0
pyproj==1.9.5.1
pystache==0.5.4
pytesseract==0.2.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
static==1.1.1
static3==0.7.0
Twisted==17.5.0
txaio==2.8.0
Unidecode==0.4.21
urllib3==1.21.1
webencodings==0.5.1
websocket-client==0.44.0
whitenoise==3.3.0
Willow==0.4
zope.interface==4.4.2
